move hash lock to propoer module and write test
